% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{Align_Posterior}
\alias{Align_Posterior}
\title{Align posterior FPC/score samples via Procrustes analysis}
\usage{
Align_Posterior(Samples, Data, B, Type, Routine = "sign")
}
\arguments{
\item{Samples, }{formated poseterior samples}

\item{Data, }{data used to fit the model}

\item{B, }{orthonormal basis matrix}

\item{Type, }{model type/structure}

\item{Routine, }{method of posterior FPC alignment, options are:
\itemize{
\item \strong{sign}: simple alignment based on cross-correlation signs
\item \strong{procrustes}: alignment based on Procrustes analysis
}}

\item{Reference, }{named list of FPC matrices against which to align. Later
entries correspond to lower levels of the hierarchy}
}
\value{
BFM_Samples object with extracted samples after alignment
}
\description{
Align posterior FPC/score samples via Procrustes analysis
}
\details{
Changes signs of FPCs/scores to align with a fixed reference point
(using Procrustes analysis to minimize the difference to the reference)
}
