% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit.R
\name{bfmm}
\alias{bfmm}
\title{Fits a Bayesian Functional (Mixed) Model with scalar predictors}
\usage{
bfmm(form, data, ID1 = NULL, ID2 = NULL, spline_basis = "B", spline_dim = 25)
}
\arguments{
\item{form, }{a formula indicating the functional response (in matrix form)
and the predictor covariates (scalar)}

\item{data, }{dataframe containing the data to be fit. Functional data
should be in matrix format.}

\item{ID1, }{column name containing subject-level identifiers (if available)}

\item{ID2, }{column name containing visit-level identifiers (if available)}

\item{spline_basis, }{type of spline basis used. Several are supported:
\itemize{
\item \strong{B}: Simple degree 3 B-splines, the default
\item \strong{Splinets}: A localized, orthonormal basis
\item \strong{OB}: Orthonormalized B-splines
\item \strong{Vec}: Vector orthonormal basis created through SVD
}}

\item{spline_dim, }{dimension of the spline basis}
}
\value{
BFM object containing posterior samples suitable for analysis
}
\description{
Fits a Bayesian Functional (Mixed) Model with scalar predictors
}
\details{
Fits a flexible Bayesian functional mixed effects model using the
FAST approach to functional PCA for representing random effects
}
\references{
Sartini, J., Zhou, X., Selvin, L., Zeger, S., & Crainiceanu, C. (in press).
Fast Bayesian Functional Principal Components Analysis.
Journal of Computational and Graphical Statistics.
}
