% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{calc_splines}
\alias{calc_splines}
\title{Calculates spline bases and corresponding penalty matrices}
\usage{
calc_splines(basis_name, basis_dim, args)
}
\arguments{
\item{basis_name, }{name of the spline basis to generate objects for}

\item{basis_dim, }{dimension of the desired spline basis}

\item{args, }{argument points along which to}
}
\value{
List containing the "Basis" and penalty matrix "Pen"
}
\description{
Calculates spline bases and corresponding penalty matrices
}
\details{
Able to generate a number of spline bases: B-splines,
orthogonalized B-splines, localized orthonormal splines, and vector
orthonormal splines.
}
