% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{cmd_extract_chains}
\alias{cmd_extract_chains}
\title{Extract by-chain samples from output of cmdstanr}
\usage{
cmd_extract_chains(fit_obj)
}
\arguments{
\item{fit_obj, }{result of calling $sample()}
}
\value{
Named array of posterior samples by chain
}
\description{
Extract by-chain samples from output of cmdstanr
}
\details{
Places posterior samples from cmdstanr into standard rstan format
}
