% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{cmd_extract_permuted}
\alias{cmd_extract_permuted}
\title{Extract permuted samples from output of cmdstanr}
\usage{
cmd_extract_permuted(fit_obj)
}
\arguments{
\item{fit_obj, }{result of calling $sample()}
}
\value{
Named list of formated posterior samples
}
\description{
Extract permuted samples from output of cmdstanr
}
\details{
Places posterior samples from cmdstanr into standard rstan format
}
