% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourise.R
\name{colourise}
\alias{colourise}
\title{Generate HTML \verb{<span>} element}
\usage{
colourise(text, colour)
}
\arguments{
\item{text}{Either a character vector, or something coercible to one.}

\item{colour}{A character vector.}
}
\value{
A character vector.
}
\description{
Vectorised over \code{text} and \code{colour}.
}
\examples{
colourise("something should be red", "red")
colourise(letters[1:3], "darkgray")
colourise(letters[1:3], c("#ed2024", "#278e43", "#febd11"))
}
