% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_chain}
\alias{extract_chain}
\title{Place posterior samples in usable format, separating the chains}
\usage{
extract_chain(Samples, Data, B_out, Type)
}
\arguments{
\item{Samples, }{samples from call to STAN, organized by-chain}

\item{Data, }{data used to fit the STAN model}

\item{B_out, }{orthonormal basis matrix}

\item{Type, }{model type/structure}
}
\value{
List of posterior draws organized by chain
}
\description{
Place posterior samples in usable format, separating the chains
}
\details{
Places posterior samples into named lists of appropriate
dimension. Samples from multiple chains are each given their own list entry.
}
