% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_permuted}
\alias{extract_permuted}
\title{Place posterior samples in usable format, interleaving the chains}
\usage{
extract_permuted(Samples, Data, B_out, Type)
}
\arguments{
\item{Samples, }{samples from call to STAN}

\item{Data, }{data used to fit the STAN model}

\item{B_out, }{orthonormal basis matrix}

\item{Type, }{model type/structure}
}
\value{
list containing the formatted samples
}
\description{
Place posterior samples in usable format, interleaving the chains
}
\details{
Places posterior samples into named lists of appropriate
dimension. Samples from multiple chains are interleaved in
random order.
}
