% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{FE_Est}
\alias{FE_Est}
\title{Produce posterior estimate of the fixed effects}
\usage{
FE_Est(W_FE, B)
}
\arguments{
\item{W_FE, }{list of fixed effects spline weight (Psi) samples}

\item{B, }{orthonormal basis matrix}
}
\value{
posterior estimate of the fixed effects in matrix form
}
\description{
Produce posterior estimate of the fixed effects
}
\details{
Estimates mean spline coefficients for the fixed effects, then
applies the basis transformation
}
