% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{FPC_Est}
\alias{FPC_Est}
\title{Produce posterior estimate of the eigenfunctions}
\usage{
FPC_Est(W, S, B, K)
}
\arguments{
\item{W, }{list of eigenfunction spline weight (Psi) samples}

\item{S, }{list of eigenfunction score samples}

\item{B, }{orthonormal basis matrix}

\item{K, }{number of principal components}
}
\value{
posterior estimate of the eigenfunctions in matrix form
}
\description{
Produce posterior estimate of the eigenfunctions
}
\details{
Estimates mean spline coefficients for each observed data function,
then leverages that the eigenvectors in the latent spline space correspond
to eigenfunctions in the data space.
}
