% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{Mat_Est}
\alias{Mat_Est}
\title{Produce posterior estimate of any matrix-structured parameters}
\usage{
Mat_Est(Var)
}
\arguments{
\item{Var, }{list of matrix variate samples}
}
\value{
posterior estimate of the matrix parameter
}
\description{
Produce posterior estimate of any matrix-structured parameters
}
\details{
Estimates posterior mean value for each matrix entry
}
