% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{Mat_Infer}
\alias{Mat_Infer}
\title{Produce posterior credible interval for any matrix-structured parameter}
\usage{
Mat_Infer(Var, Alpha)
}
\arguments{
\item{Var, }{list of matrix variate samples}

\item{Alpha, }{alpha level (type 1 error) for the intervals}
}
\value{
named list of credible interval bounds in matrix format
}
\description{
Produce posterior credible interval for any matrix-structured parameter
}
\details{
Estimates posterior credible intervals for each matrix entry
(point-wise)
}
