% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{model_input}
\alias{model_input}
\title{Formats input data and constants for Bayesian Functional (Mixed) Models}
\usage{
model_input(
  form,
  data,
  id,
  visit,
  K1,
  K2,
  spline_basis,
  spline_dim,
  args,
  out_args
)
}
\arguments{
\item{form, }{a formula indicating the functional response (in matrix form)
and the predictor covariates (scalar)}

\item{data, }{dataframe containing the data to be fit. Functional data
should be in matrix format.}

\item{id, }{column name containing subject-level identifiers (if available)}

\item{visit, }{column name containing visit-level identifiers (if available)}

\item{K1, }{number of FPCs used at the first level, default chosen empirically
to account for 95\% variability}

\item{K2, }{number of FPCs used at the second level, default chosen empirically
to account for 95\% variability}

\item{spline_basis, }{type of spline basis used. Several are supported:
\itemize{
\item \strong{B}: Simple degree 3 B-splines, the default
\item \strong{Splinets}: A localized, orthonormal basis
\item \strong{OB}: Orthonormalized B-splines
\item \strong{Vec}: Vector orthonormal basis created through SVD
}}

\item{spline_dim, }{dimension of the spline basis, defaults to 25}

\item{args, }{argument points at which the function is evaluated, defaults to
equally-spaced points between 0 and 1}

\item{out_args, }{output argument points at which functions are evaluated, defaults
to args}
}
\value{
list of formatted input data and model type labels
}
\description{
Formats input data and constants for Bayesian Functional (Mixed) Models
}
\details{
Places all inputs in the necessary format and record model
structure
}
