% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{model_output}
\alias{model_output}
\title{Formats model output of Bayesian Functional (Mixed) Model fits to be
provided to the user}
\usage{
model_output(Estimates, Inferences, Type, VarNames)
}
\arguments{
\item{Estimates, }{list of posterior mean estimates}

\item{Inferences, }{list of posterior credible intervals}

\item{Type, }{model type/structure}

\item{VarNames, }{ordered names of predictor covariates}
}
\value{
list of formatted input data and model type labels
}
\description{
Formats model output of Bayesian Functional (Mixed) Model fits to be
provided to the user
}
\details{
Places all outputs in readable, accessible formats
}
