% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{\%out\%}
\alias{\%out\%}
\title{Find unmatched values}
\usage{
x \%out\% table
}
\arguments{
\item{x}{A vector. Values to be unmatched.}

\item{table}{A vector. Values to be unmatched against.}
}
\value{
A logical vector.
}
\description{
\%out\% is an alternative way of using \%in\% for unmatched values,
which provides more similarity with SQL syntax. Returns a logical vector
indicating if there is an unmatch or not for its left operand.
}
\examples{
1:10 \%out\% 1:3
letters \%out\% c("a", "e", "i", "o", "u")
}
\seealso{
\code{\link[=match]{match()}}
}
