% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{pen_numeric}
\alias{pen_numeric}
\title{Numerical integration to derive penalty matrices}
\usage{
pen_numeric(d2, basis_dim, upper = 1)
}
\arguments{
\item{d2, }{evaluated second derivatives of the basis}

\item{basis_dim, }{dimension of the desired spline basis}

\item{upper, }{upper extent of the evaluation integral}
}
\value{
Outer product of second derivative functions for penalty
}
\description{
Numerical integration to derive penalty matrices
}
\details{
Uses numerical integration techniques to calculate penalty
matrices for arbitrary basis functions
}
