% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{Posterior_Estimates}
\alias{Posterior_Estimates}
\title{Calculate posterior estimates for all components of interest}
\usage{
Posterior_Estimates(Samples, B, Data, Type)
}
\arguments{
\item{Samples, }{aligned and formated posterior samples}

\item{B, }{orthonormal basis matrix}

\item{Data, }{data used to fit the model}

\item{Type, }{model type/structure}
}
\value{
named list of posterior estimates mirroring the output of extract
}
\description{
Calculate posterior estimates for all components of interest
}
\details{
Calculates posterior mean estimates from drawn samples
}
