% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{Posterior_Infer}
\alias{Posterior_Infer}
\title{Calculate posterior credible intervals for all components of interest}
\usage{
Posterior_Infer(Samples, B, Data, Type, Alpha = 0.05)
}
\arguments{
\item{Samples, }{aligned and formated posterior samples}

\item{B, }{orthonormal basis matrix}

\item{Data, }{data used to fit the model}

\item{Type, }{model type/structure}

\item{Alpha, }{alpha level (type 1 error) for the intervals}
}
\value{
named list of posterior credible interval quantities, mirrors the
output of extract and Posterior_Estimates
}
\description{
Calculate posterior credible intervals for all components of interest
}
\details{
Calculates credible intervals using quantiles of drawn samples
}
