% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{RHat_FE}
\alias{RHat_FE}
\title{Calculate R-Hat statistics for fixed effects functions (median and maximum
over the domain)}
\usage{
RHat_FE(FE, Data, VarNames)
}
\arguments{
\item{FE, }{fixed effect posterior samples organized by-chain}

\item{Data, }{data used to fit the model}

\item{VarNames, }{variable names for the fixed effects}
}
\value{
named vector of R-Hat statistic summaries
}
\description{
Calculate R-Hat statistics for fixed effects functions (median and maximum
over the domain)
}
\details{
Calculates Gelman-Rubin R-Hat statistics at all points along
the functional domain, summarizing by the median and maximum observed values
}
