% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{RHat_FPC_Score}
\alias{RHat_FPC_Score}
\title{Calculate R-Hat statistics for eigenfunctions (median and maximum
over the domain) and scores after alignment}
\usage{
RHat_FPC_Score(EF, Scores, EF_Estimate, Data, Level, Routine)
}
\arguments{
\item{EF, }{eigenfunction posterior samples organized by-chain}

\item{Scores, }{list of score samples}

\item{EF_Estimate, }{matrix FPC estimate}

\item{Data, }{data used to fit the model}

\item{Level, }{level of the hierarchy to consider}

\item{Routine, }{method of posterior FPC alignment, options are:
\itemize{
\item \strong{sign}: simple alignment based on cross-correlation signs
\item \strong{procrustes}: alignment based on Procrustes analysis
}}
}
\value{
vector of R-Hat statistic summaries
}
\description{
Calculate R-Hat statistics for eigenfunctions (median and maximum
over the domain) and scores after alignment
}
\details{
Calculates Gelman-Rubin R-Hat statistics for eigenfunctions at all
points along the functional domain, and for all scores. Summarizes by the
median and maximum observed values within each FPC
}
