% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{RHat_FullMod}
\alias{RHat_FullMod}
\title{Calculate R-Hat statistics for parameters within the MCMC model (fixed
effects and any present FPCs/scores)}
\usage{
RHat_FullMod(Samples, Estimates, Data, VarNames, Type, Routine = "sign")
}
\arguments{
\item{Samples, }{formatted posterior samples organized by chain}

\item{Estimates, }{object containing posterior estimates}

\item{Data, }{data used to fit the model}

\item{VarNames, }{variable names for the fixed effects}

\item{Type, }{model type/structure}

\item{Routine, }{method of posterior FPC alignment, options are:
\itemize{
\item \strong{sign}: simple alignment based on cross-correlation signs
\item \strong{procrustes}: alignment based on Procrustes analysis
}}
}
\value{
vector of R-Hat statistic summaries
}
\description{
Calculate R-Hat statistics for parameters within the MCMC model (fixed
effects and any present FPCs/scores)
}
\details{
Calculates Gelman-Rubin R-Hat statistics for functional
components at all points along the functional domain, and every score.
Summarizes by the median and maximum observed values within function
}
