% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{sign_al}
\alias{sign_al}
\title{Align posterior FPC/score samples via sign flips}
\usage{
sign_al(FPCs, Scores, Reference, Weights = NULL)
}
\arguments{
\item{FPCs, }{list of eigenfunction samples}

\item{Scores, }{list of score samples}

\item{Reference, }{FPC matrix (# observations x # FPCs)}

\item{Weights, }{list of eigenfunction spline weight samples, not required}
}
\value{
Named list with aligned FPC/Score samples
}
\description{
Align posterior FPC/score samples via sign flips
}
\details{
Changes signs of FPCs/scores to align with a fixed reference point
(according to cross-correlation between FPCs)
}
