% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{Smooth_Est}
\alias{Smooth_Est}
\title{Calculate posterior estimates for latent smooth functions}
\usage{
Smooth_Est(Samples, B, Data, Type)
}
\arguments{
\item{Samples, }{aligned and formated posterior samples}

\item{B, }{orthonormal basis matrix}

\item{Data, }{data used to fit the model}

\item{Type, }{model type/structure}
}
\value{
#participants x #observations matrix of posterior mean smooth estimates
}
\description{
Calculate posterior estimates for latent smooth functions
}
\details{
Calculates posterior mean estimates using FPCA component samples
}
