% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{Smooth_Infer}
\alias{Smooth_Infer}
\title{Calculate posterior credible intervals for latent smooth functions}
\usage{
Smooth_Infer(Samples, B, Data, Type, Alpha)
}
\arguments{
\item{Samples, }{aligned and formated posterior samples}

\item{B, }{orthonormal basis matrix}

\item{Data, }{data used to fit the model}

\item{Type, }{model type/structure}

\item{Alpha, }{alpha level (type 1 error) for the intervals}
}
\value{
named list of credible interval bounds in matrix format (# subjects
x # observations)
}
\description{
Calculate posterior credible intervals for latent smooth functions
}
\details{
Calculates posterior credible intervals using FPCA component samples
}
