% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{Splinets_basis}
\alias{Splinets_basis}
\title{Generate Splinet basis augmented with intercept and slope}
\usage{
Splinets_basis(basis_dim)
}
\arguments{
\item{basis_dim, }{dimension of the desired spline basis}
}
\value{
List containing the basis (B - list of functions) and
inner products between intercept/slope and splinets (cInt, cSlo)
}
\description{
Generate Splinet basis augmented with intercept and slope
}
\details{
Generates orthonormalized B-spline basis which maintains locality
of the standard B-splines, assisting in numerical stability and efficiency
}
