% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{Splinets_deriv2}
\alias{Splinets_deriv2}
\title{Generate second derivatives of Splinet basis augmented with intercept
and slope}
\usage{
Splinets_deriv2(basis_dim, cInt, cSlo, Upper = 1)
}
\arguments{
\item{basis_dim, }{dimension of the desired spline basis}

\item{cInt, }{inner products between splinets and intercept}

\item{cSlo, }{inner products between splinets and slope}

\item{Upper, }{upper bound of splines, can be increased for
numerical stability}
}
\value{
List of functions containing the second derivatives of basis
}
\description{
Generate second derivatives of Splinet basis augmented with intercept
and slope
}
\details{
Generates second derviatives of orthonormalized B-spline basis
which maintains locality of the standard B-splines
}
