% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_chop.R
\name{str_chop}
\alias{str_chop}
\title{Chop strings by given length}
\usage{
str_chop(string, length = 1)
}
\arguments{
\item{string}{A string in vector or list format.}

\item{length}{Length of string pieces}
}
\value{
A list.
}
\description{
str_chop() lets you chop strings into pieces by fixed length.
}
\examples{
x <- paste(letters, collapse = "")
str_chop(x)
str_chop(x, 2)
str_chop(x, 6)
}
