% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{Var_Exp}
\alias{Var_Exp}
\title{Calculate estimates of variability explained by model component}
\usage{
Var_Exp(Estimates, Data, Type)
}
\arguments{
\item{Data, }{data used to fit the model}

\item{Type, }{model type/structure}

\item{Samples, }{aligned and formated posterior samples}

\item{B, }{orthonormal basis matrix}
}
\value{
named list of variability explained by component
}
\description{
Calculate estimates of variability explained by model component
}
\details{
Calculate variability explained over the entire domain using the
standard R2 technique, separately by model component
}
