% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfmm.R
\name{bfmm}
\alias{bfmm}
\title{Fits a Bayesian Functional (Mixed) Model with scalar predictors}
\usage{
bfmm(
  form,
  data,
  alpha = NULL,
  id = NULL,
  visit = NULL,
  K1 = NULL,
  K2 = NULL,
  spline_basis = "Vec",
  spline_dim = 25,
  args = NULL,
  out_args = NULL,
  method = "MCMC",
  n_iter = 2000,
  n_burnin = floor(n_iter/2),
  n_chains = 1,
  n_cores = 1,
  treedepth = NULL,
  adapt_delta = NULL
)
}
\arguments{
\item{form, }{a formula indicating the functional response (in matrix form)
and the predictor covariates (scalar)}

\item{data, }{dataframe containing the data to be fit. Functional data
should be in matrix format.}

\item{alpha, }{alpha level (type 1 error rate) for posterior inferences,
if NULL, inference is not automatically performed}

\item{id, }{column name containing subject-level identifiers (if available)}

\item{visit, }{column name containing visit-level identifiers (if available)}

\item{K1, }{number of FPCs used at the first level, default chosen empirically
to account for 95\% variability}

\item{K2, }{number of FPCs used at the second level, default chosen empirically
to account for 95\% variability}

\item{spline_basis, }{type of spline basis used. Several are supported:
\itemize{
\item \strong{B}: Simple degree 3 B-splines, the default
\item \strong{Splinets}: A localized, orthonormal basis
\item \strong{OB}: Orthonormalized B-splines
\item \strong{Vec}: Vector orthonormal basis created through SVD
}}

\item{spline_dim, }{dimension of the spline basis, defaults to 25}

\item{args, }{argument points at which the function is evaluated, defaults to
equally-spaced points between 0 and 1}

\item{out_args, }{output argument points at which functions are evaluated, defaults
to args}

\item{method, }{Bayesian computation algorithm used, options are:
\itemize{
\item \strong{MCMC}: Markov chain Monte Carlo simulation
\item \strong{MF}: Mean field variational approximation (EXPERIMENTAL)
}}

\item{n_iter, }{number of total sampling iterations for MCMC, defaults to 2000}

\item{n_burnin, }{number of burn-in samples for MCMC, defaults to n_iter/2}

\item{n_chains, }{number of chains for MCMC, defaults to 1}

\item{n_cores, }{number of machine cores to use when running MCMC, defaults to n_chains}

\item{treedepth, }{maximum tree depth for HMC sampler}

\item{adapt_delta, }{adaptation target acceptance for HMC sampler}
}
\value{
named list of outputs:
\itemize{
\item Samples: posterior samples after applying alignment
\item Estimates: posterior estimates of all model parameters
\item Inferences: posterior credible intervals (if alpha is supplied)
\item Diagnostics: list of model diagnostics (variance explained, RHats for MCMC)
}

Outputs follow the same sub-object structure:
\itemize{
\item FE: Fixed effects (beta) matrix
\item EF(1/2): Eigenfunctions (FPCs) at levels 1 and 2
\item S(1/2): Eigenfunction scores at levels 1 and 2
\item EV(1/2): Eigenvalues at levels 1 and 2
}
}
\description{
Fits a Bayesian Functional (Mixed) Model with scalar predictors
}
\details{
Fits a flexible Bayesian functional mixed effects model using the
FAST approach to functional PCA for representing random effects
}
\references{
Sartini, J., Zhou, X., Selvin, L., Zeger, S., & Crainiceanu, C. (2025).
Fast Bayesian Functional Principal Components Analysis.
Journal of Computational and Graphical Statistics.
}
